+++ examples/intel/url/reputation.go
@@ -8,29 +8,29 @@ import (
 	"os"
 
 	"github.com/pangeacyber/pangea-go/pangea-sdk/pangea"
-	"github.com/pangeacyber/pangea-go/pangea-sdk/service/domain_intel"
+	"github.com/pangeacyber/pangea-go/pangea-sdk/service/url_intel"
 )
 
 func main() {
-	token := os.Getenv("PANGEA_DOMAIN_INTEL_TOKEN")
+	token := os.Getenv("PANGEA_URL_INTEL_TOKEN")
 	if token == "" {
 		log.Fatal("Unauthorized: No token present")
 	}
 
-	intelcli := domain_intel.New(&pangea.Config{
+	intelcli := url_intel.New(&pangea.Config{
 		Token:  token,
 		Domain: os.Getenv("PANGEA_DOMAIN"),
 	})
 
 	ctx := context.Background()
-	input := &domain_intel.DomainLookupInput{
-		Domain:   "737updatesboeing.com",
+	input := &url_intel.UrlReputationRequest{
+		Url:      "http://113.235.101.11:54384",
 		Raw:      true,
 		Verbose:  true,
-		Provider: "domaintools",
+		Provider: "crowdstrike",
 	}
 
-	response, err := intelcli.Lookup(ctx, input)
+	response, err := intelcli.Reputation(ctx, input)
 	if err != nil {
 		log.Fatal(err)
 	}