+++ pangea-sdk/internal/pangeautil/pangeautil.go
@@ -10,25 +10,40 @@ import (
 	"time"
 )
 
-func CanonicalizeStruct(v interface{}) []byte {
+func CanonicalizeStruct(v interface{}) ([]byte, error) {
 	var smap map[string]interface{}
-	ebytes, _ := json.Marshal(v)
+	ebytes, err := json.Marshal(v)
+	if err != nil {
+		return nil, err
+	}
 	// Order keys
 	json.Unmarshal(ebytes, &smap)
-	mbytes, _ := json.Marshal(smap)
-	return mbytes
+	mbytes, err := json.Marshal(smap)
+	if err != nil {
+		return nil, err
+	}
+	return mbytes, nil
 }
 
 type PangeaTimestamp time.Time
 
-const ptLayout = "2006-01-02T15:04:05.000000Z"
+const ptLayout_Z = "2006-01-02T15:04:05.000000Z"
+const ptLayout_noZ = "2006-01-02T15:04:05.000000"
 
 // UnmarshalJSON Parses the json string in the custom format
 func (ct *PangeaTimestamp) UnmarshalJSON(b []byte) (err error) {
 	s := strings.Trim(string(b), `"`)
-	nt, err := time.Parse(ptLayout, s)
-	*ct = PangeaTimestamp(nt)
-	return
+	nt, err := time.Parse(ptLayout_Z, s)
+	if err == nil {
+		*ct = PangeaTimestamp(nt)
+		return nil
+	}
+	nt, err = time.Parse(ptLayout_noZ, s)
+	if err == nil {
+		*ct = PangeaTimestamp(nt)
+		return
+	}
+	return err
 }
 
 // MarshalJSON writes a quoted string in the custom format
@@ -40,7 +55,7 @@ func (ct PangeaTimestamp) MarshalJSON() ([]byte, error) {
 func (ct *PangeaTimestamp) String() string {
 	t := time.Time(*ct)
 	utc := t.UTC()
-	return fmt.Sprintf("%q", utc.Format(ptLayout))
+	return fmt.Sprintf("%q", utc.Format(ptLayout_Z))
 }
 
 // CanonicalizeJSONMarshall is not a true canoni